refactor: moving under (page) dir
