+++ apps/web/app/(pages)/discover/search/page.tsx
@@ -2,16 +2,16 @@
 
 import { useRouter, useSearchParams } from 'next/navigation';
 import { useState, useEffect, useCallback, useMemo } from 'react';
-import { SearchFormValues } from '../../_types/search';
-import Header from "../../_components/header";
-import Footer from "../../_components/footer";
-import SearchBar from "../../_components/searchbar";
-import Masonry from "../../_components/masonry";
-import { ImageItem } from '../../_types/images';
-import { useSwipeGesture } from '../../_hooks/useSwipeGesture';
-import FilterTag from '../../_components/filterTag';
-import SearchDrawer from '../../_components/searchDrawer';
-import TextBtn from '../../_components/textBtn';
+import { SearchFormValues } from '../../../_types/search';
+import Header from "../../../_components/header";
+import Footer from "../../../_components/footer";
+import SearchBar from "../../../_components/searchbar";
+import Masonry from "../../../_components/masonry";
+import { ImageItem } from '../../../_types/images';
+import { useSwipeGesture } from '../../../_hooks/useSwipeGesture';
+import FilterTag from '../../../_components/filterTag';
+import SearchDrawer from '../../../_components/searchDrawer';
+import TextBtn from '../../../_components/textBtn';
 
 
 export default function SearchResultsPage() {
