+++ apps/web/app/(pages)/survey/2/page.tsx
@@ -1,10 +1,10 @@
 "use client";
 
 import { useRouter } from 'next/navigation';
-import Progressbar from '../../_components/progressbar';
+import Progressbar from '../../../_components/progressbar';
 import SelectionCard from '../_components/selectionCard';
 import { useSurvey } from '../SurveyContext';
-import Button from '../../_components/Button';
+import Button from '../../../_components/Button';
 import { useEffect } from 'react';
 
 import { 
@@ -15,7 +15,7 @@ import {
   ChartNoAxesColumnIncreasing as PopularityIcon,
   MapPin as LocationIcon,
 } from 'lucide-react';
-import TextBtn from '../../_components/textBtn';
+import TextBtn from '../../../_components/textBtn';
 
 export default function Page() {
   const router = useRouter();
